+++ logs/log_2022_5_27.log
@@ -1,0 +1,6 @@
+info: Jun-27-2022 20:07:59: 	{
+	"_id": "62b86aac9d2870c167b69e9d",
+	"name": "EKO",
+	"email": "aryejfa@gmail.com",
+	"status": "Login user"
+}
