+++ utils/Upload.js
@@ -1,0 +1,12 @@
+const multer = require("multer");
+const storage = multer.diskStorage({
+  destination: (req, file, cb) => {
+    cb(null, "./public/uploads");
+  },
+  filename: (req, file, cb) => {
+    cb(null, file.fieldname + "-" + Date.now() + ".jpg");
+  },
+});
+const upload = multer({ storage: storage });
+
+module.exports = upload;
